requirements.txt
opencv-contrib-python
numpy
pandas
pyttsx3
Pillow
SpeechRecognition
pyaudio
